+++ store/store_test.go
@@ -1,6 +1,7 @@
 package store
 
 import (
+	"context"
 	"encoding/json"
 	"fmt"
 	"net/url"
@@ -9,31 +10,23 @@ import (
 	"testing"
 	"time"
 
-	"github.com/globalsign/mgo"
-	"github.com/globalsign/mgo/bson"
+	"go.mongodb.org/mongo-driver/bson"
 
-	"github.com/tidepool-org/go-common/clients/mongo"
+	tpMongo "github.com/tidepool-org/go-common/clients/mongo"
 )
 
-var testingConfig = &mongo.Config{ConnectionString: "mongodb://localhost/data_test"}
+var testingConfig = &tpMongo.Config{ConnectionString: "mongodb://127.0.0.1/data_test", Database: "data_test"}
 
 func before(t *testing.T, docs ...interface{}) *MongoStoreClient {
 
 	store := NewMongoStoreClient(testingConfig)
 
 	//INIT THE TEST - we use a clean copy of the collection before we start
-	cpy := store.session.Copy()
-	defer cpy.Close()
-
 	//just drop and don't worry about any errors
-	mgoDataCollection(cpy).DropCollection()
-
-	if err := mgoDataCollection(cpy).Create(&mgo.CollectionInfo{}); err != nil {
-		t.Error("We couldn't created the deviceData collection for these tests ", err)
-	}
+	dataCollection(store).Drop(context.TODO())
 
 	if len(docs) > 0 {
-		if err := mgoDataCollection(cpy).Insert(docs...); err != nil {
+		if _, err := dataCollection(store).InsertMany(store.context, docs); err != nil {
 			t.Error("Unable to insert documents", err)
 		}
 	}
@@ -78,7 +71,7 @@ func contains(s []string, e string) bool {
 
 func basicQuery() bson.M {
 	qParams := &Params{
-		UserId:        "abc123",
+		UserID:        "abc123",
 		SchemaVersion: &SchemaVersion{Maximum: 2, Minimum: 0},
 		Dexcom:        true,
 		Medtronic:     true,
@@ -92,8 +85,8 @@ func allParams() *Params {
 	latestDataTime, _ := time.Parse(time.RFC3339, "2016-12-13T02:00:00Z")
 
 	return &Params{
-		UserId:        "abc123",
-		DeviceId:      "device123",
+		UserID:        "abc123",
+		DeviceID:      "device123",
 		SchemaVersion: &SchemaVersion{Maximum: 2, Minimum: 0},
 		Date:          Date{"2015-10-07T15:00:00.000Z", "2015-10-11T15:00:00.000Z"},
 		Types:         []string{"smbg", "cbg"},
@@ -116,16 +109,16 @@ func allParamsQuery() bson.M {
 	return generateMongoQuery(allParams())
 }
 
-func allParamsIncludingUploadIdQuery() bson.M {
+func allParamsIncludingUploadIDQuery() bson.M {
 	qParams := allParams()
-	qParams.UploadId = "xyz123"
+	qParams.UploadID = "xyz123"
 
 	return generateMongoQuery(qParams)
 }
 
 func typeAndSubtypeQuery() bson.M {
 	qParams := &Params{
-		UserId:             "abc123",
+		UserID:             "abc123",
 		SchemaVersion:      &SchemaVersion{Maximum: 2, Minimum: 0},
 		Types:              []string{"smbg", "cbg"},
 		SubTypes:           []string{"stuff"},
@@ -137,18 +130,18 @@ func typeAndSubtypeQuery() bson.M {
 	return generateMongoQuery(qParams)
 }
 
-func uploadIdQuery() bson.M {
+func uploadIDQuery() bson.M {
 	qParams := &Params{
-		UserId:        "abc123",
+		UserID:        "abc123",
 		SchemaVersion: &SchemaVersion{Maximum: 2, Minimum: 0},
-		UploadId:      "xyz123",
+		UploadID:      "xyz123",
 	}
 	return generateMongoQuery(qParams)
 }
 
 func blipQuery() bson.M {
 	qParams := &Params{
-		UserId:        "abc123",
+		UserID:        "abc123",
 		SchemaVersion: &SchemaVersion{Maximum: 2, Minimum: 0},
 		LevelFilter:   []int{1, 2},
 		Date:          Date{"2015-10-07T15:00:00.000Z", "2015-11-07T15:00:00.000Z"},
@@ -159,7 +152,7 @@ func blipQuery() bson.M {
 
 func typesWithDeviceEventQuery() bson.M {
 	qParams := &Params{
-		UserId:        "abc123",
+		UserID:        "abc123",
 		SchemaVersion: &SchemaVersion{Maximum: 2, Minimum: 0},
 		LevelFilter:   []int{1, 2},
 		Date:          Date{"2015-10-07T15:00:00.000Z", "2015-11-07T15:00:00.000Z"},
@@ -171,7 +164,7 @@ func typesWithDeviceEventQuery() bson.M {
 
 func typesWithoutDeviceEventQuery() bson.M {
 	qParams := &Params{
-		UserId:        "abc123",
+		UserID:        "abc123",
 		SchemaVersion: &SchemaVersion{Maximum: 2, Minimum: 0},
 		LevelFilter:   []int{1, 2},
 		Date:          Date{"2015-10-07T15:00:00.000Z", "2015-11-07T15:00:00.000Z"},
@@ -183,7 +176,7 @@ func typesWithoutDeviceEventQuery() bson.M {
 
 func typesWithDeviceEventAndSubTypeQuery() bson.M {
 	qParams := &Params{
-		UserId:        "abc123",
+		UserID:        "abc123",
 		SchemaVersion: &SchemaVersion{Maximum: 2, Minimum: 0},
 		LevelFilter:   []int{1, 2},
 		Date:          Date{"2015-10-07T15:00:00.000Z", "2015-11-07T15:00:00.000Z"},
@@ -199,7 +192,7 @@ func testDataForLatestTests() map[string]bson.M {
 		"upload1": bson.M{
 			"_active":        true,
 			"_userId":        "abc123",
-			"_schemaVersion": 1,
+			"_schemaVersion": int32(1),
 			"time":           "2019-03-15T01:24:28.000Z",
 			"type":           "upload",
 			"deviceId":       "dev123",
@@ -208,7 +201,7 @@ func testDataForLatestTests() map[string]bson.M {
 		"cbg1": bson.M{
 			"_active":        true,
 			"_userId":        "abc123",
-			"_schemaVersion": 1,
+			"_schemaVersion": int32(1),
 			"time":           "2019-03-15T00:42:51.902Z",
 			"type":           "cbg",
 			"units":          "mmol/L",
@@ -219,7 +212,7 @@ func testDataForLatestTests() map[string]bson.M {
 		"upload2": bson.M{
 			"_active":        true,
 			"_userId":        "abc123",
-			"_schemaVersion": 1,
+			"_schemaVersion": int32(1),
 			"time":           "2019-03-14T01:24:28.000Z",
 			"type":           "upload",
 			"deviceId":       "dev456",
@@ -228,7 +221,7 @@ func testDataForLatestTests() map[string]bson.M {
 		"cbg2": bson.M{
 			"_active":        true,
 			"_userId":        "abc123",
-			"_schemaVersion": 1,
+			"_schemaVersion": int32(1),
 			"time":           "2019-03-14T00:42:51.902Z",
 			"type":           "cbg",
 			"units":          "mmol/L",
@@ -239,7 +232,7 @@ func testDataForLatestTests() map[string]bson.M {
 		"upload3": bson.M{
 			"_active":        true,
 			"_userId":        "xyz123",
-			"_schemaVersion": 1,
+			"_schemaVersion": int32(1),
 			"time":           "2019-03-19T01:24:28.000Z",
 			"type":           "upload",
 			"deviceId":       "dev789",
@@ -248,7 +241,7 @@ func testDataForLatestTests() map[string]bson.M {
 		"cbg3": bson.M{
 			"_active":        true,
 			"_userId":        "xyz123",
-			"_schemaVersion": 1,
+			"_schemaVersion": int32(1),
 			"time":           "2019-03-19T00:42:51.902Z",
 			"type":           "cbg",
 			"units":          "mmol/L",
@@ -332,7 +325,7 @@ func TestStore_generateMongoQuery_allParams(t *testing.T) {
 
 func TestStore_generateMongoQuery_allparamsWithUploadId(t *testing.T) {
 
-	query := allParamsIncludingUploadIdQuery()
+	query := allParamsIncludingUploadIDQuery()
 
 	expectedQuery := bson.M{
 		"_userId":        "abc123",
@@ -355,7 +348,7 @@ func TestStore_generateMongoQuery_allparamsWithUploadId(t *testing.T) {
 
 func TestStore_generateMongoQuery_uploadId(t *testing.T) {
 
-	query := uploadIdQuery()
+	query := uploadIDQuery()
 
 	expectedQuery := bson.M{
 		"_userId":        "abc123",
@@ -582,7 +575,7 @@ func TestStore_GetParams_Empty(t *testing.T) {
 	schema := &SchemaVersion{Minimum: 1, Maximum: 3}
 
 	expectedParams := &Params{
-		UserId:        "1122334455",
+		UserID:        "1122334455",
 		SchemaVersion: schema,
 		Types:         []string{""},
 		SubTypes:      []string{""},
@@ -607,7 +600,7 @@ func TestStore_GetParams_Medtronic(t *testing.T) {
 	schema := &SchemaVersion{Minimum: 1, Maximum: 3}
 
 	expectedParams := &Params{
-		UserId:        "1122334455",
+		UserID:        "1122334455",
 		SchemaVersion: schema,
 		Types:         []string{""},
 		SubTypes:      []string{""},
@@ -633,11 +626,11 @@ func TestStore_GetParams_UploadId(t *testing.T) {
 	schema := &SchemaVersion{Minimum: 1, Maximum: 3}
 
 	expectedParams := &Params{
-		UserId:        "1122334455",
+		UserID:        "1122334455",
 		SchemaVersion: schema,
 		Types:         []string{""},
 		SubTypes:      []string{""},
-		UploadId:      "xyz123",
+		UploadID:      "xyz123",
 		LevelFilter:   []int{0, 1},
 	}
 
@@ -1206,16 +1199,24 @@ func TestStore_LatestNoFilter(t *testing.T) {
 	store := before(t, storeData...)
 
 	qParams := &Params{
-		UserId:        "abc123",
+		UserID:        "abc123",
 		SchemaVersion: &SchemaVersion{Maximum: 2, Minimum: 0},
 		Latest:        true,
 	}
 
-	var result bson.M
-	iter := store.GetDeviceData(qParams)
+	iter, err := store.GetDeviceData(qParams)
+	if err != nil {
+		t.Error("Error querying Mongo")
+	}
+
 	resultCount := 0
 	processedResultCount := 0
-	for iter.Next(&result) {
+	for iter.Next(store.context) {
+		var result bson.M
+		err := iter.Decode(&result)
+		if err != nil {
+			t.Error("Mongo Decode error")
+		}
 		// For `latest`, we need to look inside the returned results at the `latest_doc` field
 		result = result["latest_doc"].(bson.M)
 		switch dataType := result["type"]; dataType {
@@ -1247,17 +1248,25 @@ func TestStore_LatestTypeFilter(t *testing.T) {
 	store := before(t, storeData...)
 
 	qParams := &Params{
-		UserId:        "abc123",
+		UserID:        "abc123",
 		SchemaVersion: &SchemaVersion{Maximum: 2, Minimum: 0},
 		Types:         []string{"cbg"},
 		Latest:        true,
 	}
 
-	var result bson.M
-	iter := store.GetDeviceData(qParams)
+	iter, err := store.GetDeviceData(qParams)
+	if err != nil {
+		t.Error("Error querying Mongo")
+	}
+
 	resultCount := 0
 	processedResultCount := 0
-	for iter.Next(&result) {
+	for iter.Next(store.context) {
+		var result bson.M
+		err := iter.Decode(&result)
+		if err != nil {
+			t.Error("Mongo Decode error")
+		}
 		// For `latest`, we need to look inside the returned results at the `latest_doc` field
 		result = result["latest_doc"].(bson.M)
 		switch dataType := result["type"]; dataType {
@@ -1283,17 +1292,25 @@ func TestStore_LatestUploadIdFilter(t *testing.T) {
 	store := before(t, storeData...)
 
 	qParams := &Params{
-		UserId:        "abc123",
+		UserID:        "abc123",
 		SchemaVersion: &SchemaVersion{Maximum: 2, Minimum: 0},
-		UploadId:      "zzz4bb16e27c4973c2f37af81784a05d",
+		UploadID:      "zzz4bb16e27c4973c2f37af81784a05d",
 		Latest:        true,
 	}
 
-	var result bson.M
-	iter := store.GetDeviceData(qParams)
+	iter, err := store.GetDeviceData(qParams)
+	if err != nil {
+		t.Error("Error querying Mongo")
+	}
+
 	resultCount := 0
 	processedResultCount := 0
-	for iter.Next(&result) {
+	for iter.Next(store.context) {
+		var result bson.M
+		err := iter.Decode(&result)
+		if err != nil {
+			t.Error("Mongo Decode error")
+		}
 		// For `latest`, we need to look inside the returned results at the `latest_doc` field
 		result = result["latest_doc"].(bson.M)
 		switch dataType := result["type"]; dataType {
@@ -1325,17 +1342,25 @@ func TestStore_LatestDeviceIdFilter(t *testing.T) {
 	store := before(t, storeData...)
 
 	qParams := &Params{
-		UserId:        "xyz123",
-		DeviceId:      "dev789",
+		UserID:        "xyz123",
+		DeviceID:      "dev789",
 		SchemaVersion: &SchemaVersion{Maximum: 2, Minimum: 0},
 		Latest:        true,
 	}
 
-	var result bson.M
-	iter := store.GetDeviceData(qParams)
+	iter, err := store.GetDeviceData(qParams)
+	if err != nil {
+		t.Error("Error querying Mongo")
+	}
+
 	resultCount := 0
 	processedResultCount := 0
-	for iter.Next(&result) {
+	for iter.Next(store.context) {
+		var result bson.M
+		err := iter.Decode(&result)
+		if err != nil {
+			t.Error("Mongo Decode error")
+		}
 		// For `latest`, we need to look inside the returned results at the `latest_doc` field
 		result = result["latest_doc"].(bson.M)
 		switch dataType := result["type"]; dataType {